[mate][fix][add] css, js 공통 적용되는 부분 수정, insert file 이미지 업로드 컨트롤러 서비스 수정, 동행인 게시글 작성 부분 공백일 경우 인서트되지 않도록 경고문 hide() show() 적용
