+++ .gitignore
@@ -152,8 +152,4 @@ $RECYCLE.BIN/
 target/*
 
 
-# user properties
-src/main/resources/properties/*
-
-
 # End of https://www.toptal.com/developers/gitignore/api/windows,intellij+all,macos
