+++ src/main/java/com/multi/hontrip/mate/service/MateService.java
@@ -1,6 +1,7 @@
 package com.multi.hontrip.mate.service;
 
 import com.multi.hontrip.mate.dto.*;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 
@@ -21,7 +22,7 @@ public interface MateService {
 
     void commentEdit(MateCommentDTO mateCommentDTO); //댓글 수정
 
-    public void insert(MateBoardInsertDTO mateBoardInsertDTO); //동행인 게시글 작성하기
+    public void insert(MultipartFile file, MateBoardInsertDTO mateBoardInsertDTO); //동행인 게시글 작성하기
 
     public MateBoardSelectOneDTO selectOne(long id);    //동행인 상세게시글 가져오기
