+++ src/main/java/com/multi/hontrip/mate/service/NotificationService.java
@@ -6,6 +6,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
+import java.io.IOException;
+
 @Service
 @RequiredArgsConstructor
 public class NotificationService {
@@ -18,4 +20,17 @@ public class NotificationService {
         String id = userId + "_" + System.currentTimeMillis();
         return new SseEmitter();
     }
+
+    private void sendToClient(SseEmitter emitter, String id, Object data) {
+        try {
+            emitter.send(SseEmitter.event()
+                    .id(id)
+                    .name("sse")
+                    .data(data));
+        } catch (IOException exception) {
+            /*emitterRepository.deleteById(id);*/
+            throw new RuntimeException("연결 오류!");
+        }
+    }
+
 }
