+++ src/main/resources/properties/user/kakao.properties
@@ -1,0 +1,6 @@
+kakao.auth.url=https://kauth.kakao.com
+kakao.client.id=
+kakao.client.secret=
+kakao.redirect.url=http://localhost:8080/hontrip/user/kakao/callback
+kakao.api.url=https://kapi.kakao.com
+kakao.logout.redirect.url=http://localhost:8080/hontrip/user/kakao/logout
