+++ src/main/resources/properties/user/naver.properties
@@ -1,0 +1,6 @@
+naver.auth.url=https://nid.naver.com
+naver.client.id=
+naver.client.secret=
+naver.redirect.url=http://localhost:8080/hontrip/user/naver/callback
+naver.api.url=https://openapi.naver.com
+naver.logout.redirect.url=http://localhost:8080/hontrip/user/kakao/logout
