+++ src/main/webapp/WEB-INF/views/layout/layout.jsp
@@ -1,5 +1,5 @@
 <%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<%@ taglib prefix="tiles" uri="http://tiles.apache.org/tags-tiles"%>
+<%@ taglib prefix="tiles" uri="http://tiles.apache.org/tags-tiles" %>
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <html>
 <head>
@@ -9,29 +9,25 @@
     <link rel="shortcut icon" href="<%= request.getContextPath() %>/resources/assets/img/favicon.png">
     <link rel="stylesheet" href="<c:url value="/resources/assets/css/plugins.css"/>">
     <link rel="stylesheet" href="<c:url value="/resources/assets/css/style.css"/>">
-    <link rel="stylesheet" href="<c:url value="/resources/css/mate_board_insert.css"/>">
-    <link rel="stylesheet" href="<c:url value="/resources/css/mate_board_selectOne.css"/>">
-    <link rel="stylesheet" href="<c:url value="/resources/css/mate_board_update.css"/>">
+    <link rel="stylesheet" href="<c:url value="/resources/css/mate.css"/>">
     <link rel="stylesheet" href="<c:url value="/resources/css/plan.css"/>">
     <link rel="stylesheet" href="<c:url value="/resources/css/user.css"/>">
-    <link rel="preload" href="<c:url value="/resources/assets/css/fonts/urbanist.css"/>"as="style" onload="this.rel='stylesheet'">
+    <link rel="preload" href="<c:url value="/resources/assets/css/fonts/urbanist.css"/>" as="style"
+          onload="this.rel='stylesheet'">
     <script type="text/javascript" src="<c:url value="/resources/assets/js/plugins.js"/>" defer></script>
     <script type="text/javascript" src="<c:url value="/resources/assets/js/theme.js"/>" defer></script>
     <script type="text/javascript" src="<c:url value="/resources/js/jquery-3.7.0.js"/>" defer></script>
-    <script type="text/javascript" src="<c:url value="/resources/js/chat.js"/>" defer></script>
-    <script type="text/javascript" src="<c:url value="/resources/js/mate_board_insert.js"/>" defer></script>
-    <script type="text/javascript" src="<c:url value="/resources/js/mate_board_selectOne.js"/>" defer></script>
-    <script type="text/javascript" src="<c:url value="/resources/js/mate_board_update.js"/>" defer></script>
+    <script type="text/javascript" src="<c:url value="/resources/js/mate.js"/>" defer></script>
 </head>
 <body>
 <div class="content-wrapper">
-    <tiles:insertAttribute name="header" />
-    <tiles:insertAttribute name="body" />
+    <tiles:insertAttribute name="header"/>
+    <tiles:insertAttribute name="body"/>
 </div>
-<tiles:insertAttribute name="footer" />
+<tiles:insertAttribute name="footer"/>
 <div class="progress-wrap">
     <svg class="progress-circle svg-content" width="100%" height="100%" viewBox="-1 -1 102 102">
-        <path d="M50,1 a49,49 0 0,1 0,98 a49,49 0 0,1 0,-98" />
+        <path d="M50,1 a49,49 0 0,1 0,98 a49,49 0 0,1 0,-98"/>
     </svg>
 </div>
 </body>
