--- src/main/webapp/resources/css/mate_board_selectOne.css
@@ -1,33 +1,0 @@
-/*body {
-           width: 700px;
-           margin: 0 auto;
-           padding: 20px;
-           color: #292929;
-           font-size: 12px;
-       }*/
-
-.modal {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    background: rgba(0, 0, 0, 0.1);
-    top: 0;
-    left: 0;
-    display: none;
-}
-
-.modal_content {
-    width: 400px;
-    height: 200px;
-    background: #FFF;
-    border-radius: 10px;
-    position: relative;
-    top: 50%;
-    left: 50%;
-    margin-top: -100px;
-    margin-left: -200px;
-    text-align: center;
-    box-sizing: border-box;
-    line-height: 55px;
-    cursor: pointer;
-}
